// DemoEnergiaMVMFrontFork/src/app/models/InfoCompraEnergia.ts
export interface InfoCompraEnergia {
    ownerCliente: string;
    dirContratoCliente: string;
    empresaCliente: string;
    dirContratoGerador: string;
    empresaGerador: string;
    dirPlanta: string;
    nombrePlanta: string;
    dirComercializador: string;
    empresaComercializador: string;
    tipoEnergia: string;
    cantidadEnergia: number;
    fechaAprobacion: string;
    fechaAprobacionNumber: number;
    index: number;
}